refactor(spectacular): merge option interfaces into the same file as their function
