+++ packages/spectacular/src/index.ts
@@ -1,16 +1,13 @@
 // Application testing API
 export * from './lib/application-testing/application-harness/create-application-harness';
-export * from './lib/application-testing/application-harness/create-application-harness-options';
 export * from './lib/application-testing/application-harness/spectacular-application-harness';
 // Feature testing API
 //   Configuration
 export * from './lib/feature-testing/configuration/provide-spectacular-feature-test';
 //   Feature harness
 export * from './lib/feature-testing/feature-harness/create-feature-harness';
-export * from './lib/feature-testing/feature-harness/create-feature-harness-options';
 export * from './lib/feature-testing/feature-harness/spectacular-feature-harness';
 //   Feature testing module
-export * from './lib/feature-testing/feature-testing-module/spectacular-feature-testing-module-options';
 export * from './lib/feature-testing/feature-testing-module/spectacular-feature-testing-root.module';
 export * from './lib/feature-testing/feature-testing-module/spectacular-feature-testing.module';
 //   Navigation
@@ -18,7 +15,6 @@ export * from './lib/feature-testing/navigation/spectacular-feature-location';
 export * from './lib/feature-testing/navigation/spectacular-feature-router';
 // Pipe testing API
 export * from './lib/pipe-testing/pipe-harness/create-pipe-harness';
-export * from './lib/pipe-testing/pipe-harness/create-pipe-harness-options';
 export * from './lib/pipe-testing/pipe-harness/spectacular-pipe-harness';
 // Shared testing API
 export * from './lib/shared/app-component/spectacular-app.component';
