--- packages/spectacular/src/lib/application-testing/application-harness/create-application-harness-options.ts
@@ -1,9 +1,0 @@
-import { NgModule } from '@angular/core';
-
-/**
- * Application harness options.
- */
-export type CreateApplicationHarnessOptions = Pick<
-  NgModule,
-  'imports' | 'providers'
->;
