--- packages/spectacular/src/lib/application-testing/util-bootstrapping/bootstrap-component-options.ts
@@ -1,8 +1,0 @@
-import { NgZone, Type } from '@angular/core';
-
-export interface BootstrapComponentOptions<TRootComponent> {
-  readonly autoDetectChanges?: boolean;
-  readonly component: Type<TRootComponent>;
-  readonly ngZone: NgZone | null;
-  readonly tag: string;
-}
