+++ packages/spectacular/src/lib/application-testing/util-bootstrapping/bootstrap-component.ts
@@ -1,8 +1,18 @@
-import { ApplicationInitStatus, ApplicationRef } from '@angular/core';
+import {
+  ApplicationInitStatus,
+  ApplicationRef,
+  NgZone,
+  Type,
+} from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-
 import { ensureFreshRootElement } from '../util-dom/ensure-fresh-root-element';
-import { BootstrapComponentOptions } from './bootstrap-component-options';
+
+export interface BootstrapComponentOptions<TRootComponent> {
+  readonly autoDetectChanges?: boolean;
+  readonly component: Type<TRootComponent>;
+  readonly ngZone: NgZone | null;
+  readonly tag: string;
+}
 
 async function waitForApplicationInitializers(): Promise<void> {
   const applicationInitStatus = TestBed.inject(ApplicationInitStatus);
