--- packages/spectacular/src/lib/feature-testing/feature-testing-module/spectacular-feature-testing-module-options.ts
@@ -1,28 +1,0 @@
-import { ExtraOptions, Routes } from '@angular/router';
-
-/**
- * Feature testing options for `SpectacularFeatureTestingModule.withFeature`.
- */
-export interface SpectacularFeatureTestingModuleOptions {
-  /**
-   * The path prefix used to load the routes of the specified Angular feature
-   * module, for example `'heroes'`.
-   */
-  readonly featurePath: string;
-  /**
-   * Optional Angular `Router` options.
-   */
-  readonly routerOptions?: ExtraOptions;
-  /**
-   * One or more feature routes to load.
-   *
-   * NOTE! It is unnecessary to lazy-load feature modules in tests, so we can
-   * statically return an Angular module from the `loadChildren` callback.
-   *
-   * @example
-   * ```typescript
-   * [{ path: 'heroes', loadChildren: () => HeroesModule }]
-   * ```
-   */
-  readonly routes: Routes;
-}
