+++ packages/spectacular/src/lib/pipe-testing/pipe-harness/create-pipe-harness.ts
@@ -1,9 +1,35 @@
+import type { NgModule, PipeTransform, Type } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import type { Observable } from 'rxjs';
 import { SpectacularPipeComponent } from '../pipe-component/spectacular-pipe.component';
-import { CreatePipeHarnessOptions } from './create-pipe-harness-options';
 import { SpectacularPipeHarness } from './spectacular-pipe-harness';
+
+/**
+ * Angular pipe harness options.
+ */
+export interface CreatePipeHarnessOptions<TValue>
+  extends Pick<NgModule, 'declarations' | 'imports' | 'providers'> {
+  /**
+   * The type of the Angular pipe-under-test, for example `CamelizePipe`.
+   */
+  readonly pipe: Type<PipeTransform>;
+  /**
+   * The name of the Angular pipe-under-test, for example `camelize`.
+   */
+  readonly pipeName: string;
+  /**
+   * The template used to test the Angular pipe, for example
+   * `'{{ value | camelize }}'`.
+   *
+   * NOTE! The `value` property is in context of the template.
+   */
+  readonly template?: string;
+  /**
+   * The initial value passed through the Angular pipe.
+   */
+  readonly value: Observable<TValue> | TValue | null;
+}
 
 function createPipeComponentTemplate(innerTemplate: string): string {
   return `<span id="${textId}">${innerTemplate}</span>`;
