--- packages/spectacular/src/lib/pipe-testing/test-util/create-common-pipe-harness-options.ts
@@ -1,6 +1,0 @@
-import type { Observable } from 'rxjs';
-
-export interface CreateCommonPipeHarnessOptions<TValue> {
-  readonly template: string;
-  readonly value: Observable<TValue> | TValue | null;
-}
